+++ src/app/components/myreviews/myreviews.component.ts
@@ -9,6 +9,7 @@ import { AuthService } from 'src/app/Services/auth.service';
 import {LiveAnnouncer} from '@angular/cdk/a11y';
 import { ReviewService } from 'src/app/Services/review.service';
 import { DialogReviewComponent } from 'src/app/dialogs/dialog-review/dialog-review.component';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-myreviews',
@@ -43,7 +44,8 @@ export class MyreviewsComponent implements AfterViewInit {
 
   openDialog() {
     this.dialog.open(DialogReviewComponent, {
-      width:"auto"
+      width:"auto",
+      maxWidth:"700px"
     }).afterClosed().subscribe(()=>
         {
         this.setReviews();
@@ -57,6 +59,7 @@ export class MyreviewsComponent implements AfterViewInit {
   editReview(element: ReviewModel){
     this.dialog.open(DialogReviewComponent, {
       width:"auto",
+      maxWidth:"700px",
       data: element   
          
     }).afterClosed().subscribe(()=>
@@ -67,13 +70,27 @@ export class MyreviewsComponent implements AfterViewInit {
   }
 
   deleteReview(element: ReviewModel){
-    this.reviewService.deleteReview(element.id).subscribe(data =>{
-      this.setReviews();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this review!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      confirmButtonColor : '#3378cc',
+      cancelButtonText: 'No, keep it',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.reviewService.deleteReview(element.id).subscribe(data =>{
+          this.setReviews();
+        });
+      } else if (result.isDismissed) {
+
+      }
     });
   }
 
   private setReviews(){
-    this.reviewService.GetReviews()
+    this.reviewService.getMtReviews()
         .subscribe(data => {
           this.data = data; 
           this.dataSource = new MatTableDataSource(this.data);
